% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/govcolumn.R
\name{govcolumn}
\alias{govcolumn}
\title{Create a simple column chart}
\usage{
govcolumn(data, x, y, series = NULL, position = "dodge")
}
\arguments{
\item{data}{A data frame for plotting}

\item{x}{The variable for the x-axis}

\item{y}{The variable for the y-axis}

\item{series}{Optional. The variable denoting the series}

\item{position}{How the series are arranged, default is "dodge", can be
any \link[ggplot2]{ggplot} position, the main alternative is "stack".}
}
\value{
A ggplot object
}
\description{
A function for creating a simple column chart
}
