% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/govuk_colours.R
\docType{data}
\name{govuk_colours}
\alias{govuk_colours}
\title{The GOV.UK colour palette}
\format{An object of class \code{character} of length 20.}
\usage{
govuk_colours
}
\value{
A named character vector of hexadecimal colour codes.
}
\description{
\code{\link{govuk_colours}} provides a vector of named
hexadecimal colour codes from the \href{
https://design-system.service.gov.uk/styles/colour/#colour-palette}{GOV.UK
colour palette}.
}
\examples{

}
\keyword{datasets}
